+++ src/main/java/org/gruppe3nskeskyfullstack/controller/LoginController.java
@@ -19,12 +19,12 @@ public class LoginController {
         this.userService=userService;
     }
 
-    @GetMapping("/login")
-    public String showLoginPage(){
-        return "login";
-    }
+//    @GetMapping("/login")
+//    public String showLoginPage(){
+//        return "login";
+//    }
 
-    @PostMapping("/")
+    @PostMapping("/tryLogin")
     public String login(@RequestParam("email") String email,
                         @RequestParam("password") String password,
                         HttpSession session){
@@ -34,7 +34,7 @@ public class LoginController {
         if(user!=null){
             session.setAttribute("user", user);
             System.out.println("Det virkede!");
-            return "redirect:/";
+            return "redirect:/userPage";
         }
         System.out.println("forkerte logindetaljer");
         return "redirect:/"; //ved fejlet login
