--- src/main/java/org/gruppe3nskeskyfullstack/controller/WishlistController.java
@@ -1,23 +1,0 @@
-package org.gruppe3nskeskyfullstack.controller;
-
-import org.gruppe3nskeskyfullstack.service.Validation.WishlistInputValidering;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-//@RestController
-public class WishlistController {
-
-    private WishlistInputValidering validation;
-
-    public WishlistController(WishlistInputValidering validation){
-        this.validation=validation;
-    }
-
-    @GetMapping("testValideringøsnkelistenavn")
-    public  String testValidering(@RequestParam String name){
-        validation.validationWishlist(name);
-        return "Gyldig ønskelistenavn "+name;
-    }
-
-}
