+++ src/main/resources/dummy.sql
@@ -1,0 +1,6 @@
+USE WishDB;
+
+insert into wishlists (userID, name) VALUES
+    ('1','lottes fødselsdag'),
+    ('1','bents fødseldag')
+
