+++ src/main/resources/templates/index.html
@@ -8,7 +8,7 @@
 </head>
 <body>
 <div class="loginBox">
-    <form class="loginForm" th:action="@{/}" method="post">
+    <form class="loginForm" th:action="@{/tryLogin}" method="post">
         <label for="emailField">Email:</label>
         <input type="email" name="email" class="loginField" id="emailField" placeholder="indtast: 'testEmail'" required>
         <label for="passwordField">Password:</label>
